Adds more config tests.
