+++ tests/resources/config/test_config.ini
@@ -1,0 +1,3 @@
+[resources]
+aquarius.url = https://custom-aqua.url
+provider.address = 0x00bd138abd70e2f00903268f3db08f2d25677c9e
